backend/requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
firebase-admin==6.4.0
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0